refactor(agent): split subagent_runner/ops.rs (1947 lines) into ops/ directory
Sub-modules: provider (lazy resolver), observer (progress bridging),
tool_source (tool surface builder), spawn (core spawn logic).
